move interaction to component
